assesment framework component moved in respective folder
